refactor: split turn_resolver entry/stub siblings for wave-7 Slice A
Extract WorldState stub and untrusted/trusted order-engine wrappers so the
kitchen-sink resolver drops under 300 lines; wrappers build TurnResolverConfig
internally while keeping distinct public entry names.
