+++ packages/colonizethis_turn/lib/src/turn/turn_resolver_world_state_stub.dart
@@ -1,0 +1,40 @@
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'turn_resolution_sequence.dart';
+
+/// Turn resolver stub (Phase 1 compatibility). Runs phase sequence; only
+/// endOfTurn advances turn number.
+WorldState resolveTurn(WorldState current) {
+  WorldState state = current;
+  for (final phase in turnResolutionSequence) {
+    state = _runWorldStatePhase(state, phase);
+  }
+  return state;
+}
+
+WorldState _runWorldStatePhase(WorldState state, TurnPhase phase) {
+  switch (phase) {
+    case TurnPhase.orders:
+    case TurnPhase.extraction:
+    case TurnPhase.richesToTreasury:
+    case TurnPhase.production:
+    case TurnPhase.consumption:
+    case TurnPhase.spyResolution:
+    case TurnPhase.research:
+    case TurnPhase.diplomacy:
+    case TurnPhase.movement:
+    case TurnPhase.minorRegimentUpgrade:
+    case TurnPhase.navalInterceptionCombat:
+    case TurnPhase.combat:
+    case TurnPhase.buildWork:
+    case TurnPhase.worldMarket:
+      return state;
+    case TurnPhase.endOfTurn:
+      return state.copyWith(
+        turnState: state.turnState.copyWith(
+          turnNumber: state.turnState.turnNumber + 1,
+          phase: TurnPhase.orders,
+        ),
+      );
+  }
+}
